Rename 'intermediate' state to 'boot'
