+++ src/app/mod.rs
@@ -1,6 +1,6 @@
 mod state;
 pub use state::*;
 
-pub mod intermediate;
+pub mod boot;
 pub mod lobby;
 pub mod paint;
