+++ src/main.rs
@@ -157,7 +157,7 @@ async fn inner_main(language: &mut Option<Language>) -> errors::Result<()> {
    let assets = Box::new(Assets::new(ui.render(), color_scheme)?);
    let socket_system = SocketSystem::new();
    *language = Some(assets.language.clone());
-   let mut app: Option<Box<dyn AppState>> = Some(intermediate::State::new(
+   let mut app: Option<Box<dyn AppState>> = Some(boot::State::new(
       cli,
       assets,
       Arc::clone(&socket_system),
